+++ building_data_utilities/__init__.py
@@ -5,8 +5,6 @@ See also https://github.com/SEED-platform/cbl-workflow/blob/main/LICENSE.md
 
 import warnings
 
-from . import utils  # noqa: F401
-
 warnings.warn(
     ("Importing 'building_data_utilities' is deprecated and will be removed in a future release."), DeprecationWarning, stacklevel=2
 )
