--- building_data_utilities/utils/__init__.py
@@ -1,4 +1,0 @@
-"""
-SEED Platform (TM), Copyright (c) Alliance for Sustainable Energy, LLC, and other contributors.
-See also https://github.com/SEED-platform/cbl-workflow/blob/main/LICENSE.md
-"""
