+++ tests/test_shp_to_geojson.py
@@ -11,7 +11,7 @@ import tempfile
 import geopandas as gpd
 from shapely.geometry import Point, Polygon
 
-from building_data_utilities.utils.shp_to_geojson import shp_to_geojson
+from building_data_utilities.shp_to_geojson import shp_to_geojson
 
 
 class TestShpToGeoJSON:
